+++ CHANGES.rst
@@ -6,13 +6,13 @@ Changelog
 ------------------
 
 - #2033 Fix blurry Barcode and QRCode in stickers
+- #2032 Fix add-on stickers not displayed in sample type admitted stickers
 - #2031 Make the "Other reasons" text area from rejection view wider
 - #2030 Fix Email address is not displayed in clients listing
 - #2029 Fix AttributeError in Manufacturer's Instruments view
 - #2025 Display full name of analyst and submitter in analyses listing
 - #2025 Fix analyst unchanged in analyses listing after worksheet reassignment
 - #2028 Fix Definition is not displayed in Reference Samples listing
-- #2027 Fix Manufacturer is not displayed in Reference Samples listing
 - #2027 Fix Manufacturer is not displayed in Reference Samples listing
 - #2026 Fix LabManager and LabClerk cannot add preservations
 - #2024 Cannot create partitions from samples in received status